Fixes #2 - ocsp.Request has no field or method Marshal
As reported in #2 it seems like the upstream `golang.org/x/crypto/ocsp`
package changed the `Request` type to no longer define a `Marshal`
method producing errors with the invocations in cache.go and server.go
respectively.

This commit takes a stab at what the intended behaviour was meant to be
and uses the `asn1` package's `Marshal` method on the `Request` object.