Adds Vagrant support.
This commit adds a Vagrantfile to support having a VM based dev env.

I've selected Ubuntu 14.04 (Trusty) as the base image, but I'm happy to
change that if anyone has preferences. I think it's a good common
denominator OS. Currently the VM is provisioned with 2GB of RAM

The Go version being installed is fixed at 1.5.3 and there is
a hardcoded SHA1 sum in the provisioning script that the release
`.tar.gz` is compared against. If there's demand for using `1.6.0` as
the default I'm happy to switch this out.

Two bootstrap scripts are run during provisioning:

1. `bootstrap/bootstrap-dev-admin.sh` - invoked as root in the VM
2. `bootstrap/bootstrap-dev.sh` - invoked as `vagrant` user in the VM

The first is primarily used to install Go and set up system wide
settings.

The second sets up a Go workspace, adds it to the `vagrant` user's
environment, and links the project src from the host machine into the
VM's gopath.

Workflow:

1.`vagrant up`
2. modify files in your local git checkout on your host machine
3. `vagrant ssh`
4. `cd workspace/src/github.com/rolandshoemaker/stapled`
5. `go test ./...`
6. etc.