Adds `stapled-checkcert` utility.
This commit adds a new binary `stapled-checkcert` to the project. This
tool serves to perform one-off OCSP requests using the stapled codebase.

The intention is to provide an alternative to `openssl ocsp`, a command
line utility used for fetching OCSP responses using OpenSSL. The OpenSSL
'utilities' are not meant for production usage, and are written in
dubious C code.

By comparison `stapled-checkcert` is Go based, requires no OpenSSL dep,
and lets us exercise the common code already in place for stapled. In
this way it can act as a "dry-run" check for certificates prior to
placing them in the monitored directory.